REVIEW: split example components
